+++ example-app/src/index.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { ImageHandlerContext } from "@feichtmedia/imagehandler-react-sdk";
 import App from "./App";
 import "./index.css";
 
@@ -8,6 +9,18 @@ const root = ReactDOM.createRoot(
 );
 root.render(
   <React.StrictMode>
-    <App />
+    <ImageHandlerContext
+      config={{
+        endpointDomain: "images.feicht.media",
+        useHttps: true,
+        // progressiveImageLoading: false,
+        defaultStyles: {
+          transparentAltText: true,
+          fullWidth: true,
+        },
+      }}
+    >
+      <App />
+    </ImageHandlerContext>
   </React.StrictMode>
 );