+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./components";
 export { addLazyLoading, removeLazyLoading } from "./utils/lazy-loading";
+export { getImgSrc, getImgSrcSet } from "./utils/user-helper";